Changed the default 32(GuardBits) to 24 since double is not exact.
There is not a great solution to this but moving 8 bits of the mantissa into the GuardBit area is a good compromise. Prior to this, we had zero mantissa bits stored in the GuardBits and this created some issues.  If we move all 32 bits into the GuardBit area then we would just be left with 53-32 GuardBits= 21 bits. A balance of 8 bits was selected but it is best if the user selects the value.