Mark `Zero` and `One` as obsolete in `BigFloat`
The `Zero` and `One` static properties in the `BigFloat` struct have been marked with the `[Obsolete]` attribute. This is because Zero and One have a specific accuracy/precision however users may expect it to be exact. and it could cause confusing when using it.  Developers are advised to use the integer values `0` and `1` or the `ZeroWithAccuracy` and `OneWithAccuracy` alternatives instead. This change is intended to encourage better practices and prepare for potential future deprecation of these properties.