Add commented-out zero check in BigFloat struct
A commented-out line was added to the `BigFloat` struct in the `BigFloatStringsAndSpans.cs` file. The line checks if the mantissa (`intVal`) has a zero with a an accuracy above the radix point. If so it would return `"0"`, but it is currently inactive. This may or may not be added to a future implementation.